+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: Add `mediaQueries` HOC.
+- Fix: Update `primary-light` color.
+- Feature: Add `modifier` prop on `TagButton` component.
 
 ## [19.5.0] - 2018-02-13